move Course type definition
